Added an "Experimental" directory under [mrnet.git]/mrnet.  This will ease my own code management.  I'm hoping this directory will act much like [mrnet.git]/mrnet/Examples, with the exception that each experiment will have its own directory with its own Makefile inside (instead of one Makefile that builds all of the experiments).  This allows the authors of the experiments to not have to track each other's work.
In the Experimental directory, I added the PigOnMRNet experiment.  It contains a Makefile that builds the 'boar' executable.  'boar' presently uses MRNet to read tuples of a relation sharded across many backends.  It concatenates the shards together in parallel and outputs them at the frontend.  This code is a dirt-simple starting point for my Pig-on-MRNet work.