add some locking to prevent topology corruption during BE attach
when lots of BEs connect to the same CP, interleaving of threads
could lead to a situation where an incomplete topology was the
last report sent, overriding previous complete topology reports
since the last sent wins. added a mutex to ensure local subtree
string is retrieved and sent in a Packet in one logical operation,
to prevent the interleaving.