fix EDT race that causes FEs to never accept connections
The Network::_local_front_end_node pointer must be set before
the EDT thread starts, or the EDT will think the local node is
not a parent and will not add the local listening socket to
the set of event fds. When the listening socket is not monitored,
no child connections will be accepted.