PROT_TOPO_UPDATE fix for lightweight BEs
since ltwt BEs don't filter, and packets containing this tag are
sent on a non-control stream, the packets were mistakenly passed
on to the user. fixed by adding a handler function for the tag.

note that this doesn't actually implement the necessary updates,
which will need to be done if we want ltwt BEs to recover from
failures