fixes for backend attach
some previous changes were lost in a bad merge, notably
the use of a name for the topology update stream instead
of the value one.

also updated the timeout sync filter to handle packets from
nodes not yet in the topology (since they're added in the
topology update stream trans filter).