avoid joining/canceling exited threads
Bob Moench from Cray reported a case where our attempt to cancel a
recv thread during a failure scenario lead to a segfault. turns out
pthread_cancel (for Linux NPTL) is not safe to call with the id of a
thread that may have already exited, since the id is really a pointer.

The fix is to have the send/recv/EDT threads reset their ids to zero
(NULL) before they exit, and check for zero ids in other code that
cancels or joins those threads.