fixes for send thread exit on failure
Previously, in PeerNode::mark_Failed, we attempted to wake up the
blocked send thread for the peer by pushing the NullPacket on the
message queue. Unfortunately, the Message::add_Packet() method
rightfully refuses to push NullPacket, as that will lead to a
segfault in Message::send(). As a result, the send thread woke up
but didn't find anything to send and thus blocked waiting again. When
the Network was shutdown, the peer for the send thread had already
been removed, so no shutdown message was given to the send thread and
it was left running until the front-end exited.

The fix is to push a real packet (doesn't really matter what it contains,
but I used the PROT_SHUTDOWN ctl message), which the send thread will
fail to send, causing it to exit. Also, added some backup
calls to mark_Failed in Network::delete_PeerNode() just in case the
recv or EDT threads forgot to mark the peer failed.

Also changed the microbench test to turn off failure recovery and
implemented better error checking so it won't hang on failure.