new BE/filter staging support for Cray XT
The Cray ALPS tool helper library can be used to stage files to
compute nodes to avoid shared file system accesses. With this commit,
we will automatically stage the back-end executable (if given), and
will additionally stage any other files (e.g., filter libraries)
that the user explicitly requests us to stage by setting either
- the "CRAY_ALPS_STAGE_FILES" attribute, or
- the "MRNET_ALPS_STAGE_FILES" environment variable
to a colon-separated set of file paths.

To take advantage of staged libraries and executables, the PATH and
LD_LIBRARY_PATH settings must include the tool helper stage directory,
so we make the necessary updates to the environment.

As a result of this commit, the ALPS apid is distributed as part
of the Network settings map, and can be accessed using the
"CRAY_ALPS_APID" key. Additionally, if the user has the process id of
the aprun process instead of an apid, the user can pass the pid
as the attribute value for the "CRAY_ALPS_APRUN_PID" key, and we'll
do the necessary lookup to get the apid.

Finally, the attribute key "apid" is deprecated in favor of
"CRAY_ALPS_APID", but the former is still allowed for backwards
compatibility.

wow, still reading huh, well then here's some minor tidbits
- we're properly using ALPS file/directory defines instead of
  hard-coded paths that could change
- the connection retry count has been upped to 15 from 10