XT startup bug fix
the argc passed to CreateNetworkIn() in SpawnCP was wrong by +1, leading
to a segfault for spawned co-located CPs.

this commit also adds exchange of the ALPS apid in PropagateTopology /
GetTopology, so that we can look for and use the staged BE executable.

Note: I tried moving the code to initialize as a CP before
the code that spawns co-located processes, but this led to hangs
during startup (likely due to our pseudo-exec of CPs -- we just fork
and then call CreateNetworkIN, rather than doing an execvp). I added
a comment that warns about moving the code before the spawns.