child failures during network init should not cause segfaults
as reported by Bob Moench, a lookup of the topology update stream
returned null when an EDT noticed a child failure and tried to report
it, presumably because the stream hadn't been created yet. the
subsequent access to the null pointer caused the CP to segfault.

also, enables the setting of the network's failure recovery mode
via the MRNET_FAILURE_RECOVERY key in the attrs map