mrnet_topgen fixes
if a host was present in both the CP and BE input host lists, there
was a chance that the BE slots would be ignored for balanced and
generic trees.

also, GenericTree::initializeTree() had a bad iterator check, it was
checking a BE list iterator versus the end of the CP list.

finally, added a usage message that says that -p should only be
used with -h.