discard port updates for backends
Since we use a wait-for-all stream to gather port updates on
rsh networks, we need the backends to initiate the update wave.
But, backends don't have listening ports, so we shouldn't pass
these fake updates all the way to the front-end for later
broadcast. We now filter out the fake updates in the topology
updata filter code.

this commit also includes a bunch of purely cosmetic cleanup