Change build process to build separate libraries for thread-safe lightweight
When the MRNet build is configured with --enable-ltwt-threadsafe, separate
thread-safe lightweight libraries will be built alongside the ones that are
not thread-safe.  These libraries are called libxplat_lightweight_r.{a,so}
and libmrnet_lightweight_r.{a,so}.  This way users may choose the lighter
libraries if they do not desire to use lightweight BEs in a multi-threaded
environment.