Fix race condition causing hang when communicating with singlecast on BEs
On systems where there is much contention for the processor, it was
possible that the threads created by CreateBackEndNode in init_BackEnd
could try to receive on a singlecast stream before the BE created
this stream.  This causes the processing of the packet to fail and,
in the singlecast test, the BE to indefinitely hang waiting for the
packet it failed on.