Protect the 'data' member of Mutex and Monitor objects from being used after delete.
Bob Moench of Cray reported an occasional crash on shutdown. Not much is known
except the place of the failure is when a Monitor tries to lock. This has been
occasionally witnessed for a couple years, and this may fix the problem.  Even
if it does not, protecting the data member from use after deletion is something
that should be done.  The Windows patch does not work, but Cray needs this patch
and Linux has gone through many stress tests with this patch.