Prevent operation on the port stream at the FE after it is deleted
There was an uncommon race where the internal port stream could
be deleted before all recv threads on the FE.  Under loaded conditions
this caused a seg fault when loading a filter function in the tree.
There was no reason to delete the stream where it was being deleted, so
this was removed, and the stream will be deleted at shutdown.