Print thread name in xplat_printf
Added capability for the thread id and thread name to be accessible through
XPlat as well as MRNet.  To change this, had to rework the implementation of
thread-specific data and the XPlat interface to it.  All MRNet uses of it
are now updated.  This change is only for regular MRNet since lightweight
MRNet does not have internal threads.