Packet byte alignment encoding patch.
This is a patch to change the way packet data/headers are encoded
to ensure that data alignment is not modified by packet encoding
in the data section. In this patch setting of the byte ordering flag
has been removed from pdrmem_create() and is now done external to
pdrmem_create(). This allows for a single location in a packet
to specify the byte ordering for the entire packet. There are two
locations where byte ordering is now set.

1) Packet header - Byte ordering flag appears at the end of the packet
header (hdr_len - 1 byte).

2) Message Header - Byte ordering flag for the message header appears
at the beginning of the header (byte 1).