Remove destruction synchronization for Monitors and Mutexes
Adding a synchronization step in the destruction of these classes and guaranteeing
safe execution in all circumstances requires memory to be leaked in each instance
of the classes. We are assuming now that any use of a class after it is destructed
is indicative of buggy behavior elsewhere in the program, and these other behaviors
will be addressed to fix these types of problems.