Check for 0 in our Thread::Join wrappers
There are times when a thread is normally exiting that MRNet will change the stored
id for that thread to 0 very soon before the thread exits.  Anytime another
thread attempts to join that thread using the stored id, it seg faulted because
it passed '0' to the low-level join function through the wrapper.  Now the wrapper
checks for this and returns the appropriate error code, avoiding a seg fault.