Add -lboost_timer back, but disable use of boost_timer.
Removing -lboost_timer was not the correct thing to do, as I did a few
commits ago.  libboost_timer does indeed exist, but we need a better way
to check for this.  Until we figure this out, I'm disabling the
possibility of using boost_timer, since it is not at all essential.