Fix for race when a stream is deleted while being used
This was brought to light by Bob Moench at Cray.  ~Stream now makes sure
that no threads are currently using any part of a stream before it gets
deleted.  The second part of this fix remedies a hang on shutdown that
the first fix uncovered.  The event detector now checks for failed
children that did not have the chance to send a shutdown ack and acks on
its behalf so shutdown can continue.  No modifications to lightweight
because lightweight is single-threaded.