Fix for a hang that can occur during startup
This does not fix ALL hangs that can occur during startup, only one
possible one.  This patch creates the ability to abort any active control
protocol and notify a ParentNode that a child failure has occurred.  It
also changes the behavior of waitfor_ControlProtocolAcks to cease
waiting and return with an error if a failure has occurred before or
during waiting.  The EDT aborts active control protocols when a child
failure has been detected.