Remove flush in ~Stream
We can't flush here because the code that calls the Stream destructor
could hold the same lock that flush Streams would need.  This has been
shown to cause deadlocks.

Fixing this and the many other use-after-delete scenarios require a
different way of obtaining a Stream...not a dumb pointer.  This isn't
acceptable for this commit (though it is being worked on), so I'm
removing this useless flush.  Some of the cases that previously caused
deadlocks will instead cause seg faults, so consider this branch
currently UNSTABLE.