Fix for hang in IntegerAddition example
In the event that the flush call in the PROT_SUM protocol of the
example failed in the BE, a PROT_EXIT message would not be sent,
whether a message was received requesting it or not.  This caused
the FE to hang waiting for all PROT_EXIT messages.

Currently, if the FE fails, this change causes the BEs to fail to
exit instead.  This problem will be targeted with the new shutdown protocol.