The new shutdown protocol
This way of shutting down the tree does not rely on MRNet-level
acknoledgment packets when shutting down.  The lack of these acks
takes away the possibility of the FE waiting for the acks indefinitely.
This is not considered stable yet, and is being committed for more extensive testing.

NOTE: THIS BREAKS LIGHTWEIGHT! Lightweight support will be added soon.