Notification of filter load failure on CP/BE Nodes.
This patch resolves an issue where the FE would not be notified
of a failure to load a filter by CP/BE nodes. This change makes
load_filterfunc synchronous (will not return until all CPs/BEs
load the filter). The FE will print a notification to the FE
error log with the node name and filter name of the filter
that failed to load. load_filterfuncs will now return the number
of filters actually successfully loaded by ALL CPs/BEs. Users
can detect failures by comparing the number of expected filters
against the number actually loaded.