Converted a global static reference to XPlat::mutex to boost::shared_ptr
This is a patch to repair an issue where an exception is being thrown on MRNet exit.
It is unsafe to use a global static reference to any class in a multi-threaded environment
when there is a chance that any thread may attempt to acces the variable when threads are exiting
(the value can be deleted by another exiting thread before the access). Always wrap global statics
as boost::shared_ptr. Anywhere this is not the case, it is almost assuredly wrong.